pyproject.toml
[project]
name = "rt-api"
version = "1.0.0"
description = "A personal API for fetching Rotten Tomatoes movie data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
